fix(javascript): remove duplicate `echoRequester` logic
