+++ templates/javascript/tests/client/method.mustache
@@ -1,1 +1,1 @@
-{{object}}{{#path}}.{{.}}{{/path}}({{{parameters}}});
+{{object}}{{#path}}.{{.}}{{/path}}({{{parameters}}})
